added base planning service class
